Move the context and the yaml writter to separate files
